add parsing of openweathermap response

+++ data/data-models/src/main/java/com/worldturtlemedia/cyclecheck/models/weather/Wind.kt
@@ -1,7 +1,7 @@
 package com.worldturtlemedia.cyclecheck.models.weather
 
 data class Wind(
-    val speed: Double,
+    val speed: Long,
     val degree: Int
 ) {
